Catch Exception When Assuming Role for Alias
It's possible that an account might be added to OKTA but permissions
might be mangled such that the roles in the account cannot be assumed
yet. Rather than throw an exception and block folks from assuming all
roles, let's hide roles that cannot be assumed, with optional logging to
let people dive a bit deeper.

Now, if we try to get the account alias and cannot assume the role, we
return None. And if the account alias function returns None, we hide
that role from the list of roles to assume by skipping it.

okta-awscli will continue to attempt to assume the role every time they
run the command, so as soon as the permissions are fixed, the role will
appear in the list.