support AWS_SHARED_CREDENTIALS_FILE
AWS_SHARED_CREDENTIALS_FILE is a documented AWS CLI environment
variable that can be used to specify the location of the
credentials file.

Support reading the value and configuring internal AwsAuth fields
with the value extracted from it.

See https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-envvars.html

In this implementation the environment variable takes precedence
over the default $HOME/.aws location, which is used as fallback
when the environment variable is not present.